backend/requirements.txt
fastapi==0.115.1
uvicorn[standard]==0.29.0
beanie==1.29.0
motor==3.4.0
sqlalchemy[asyncio]==2.0.41
aiosqlite==0.20.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2 
python-multipart==0.0.9
email-validator